requirements.txt
streamlit
scikit-learn
joblib
pandas
numpy
requests
datasets
python-dotenv
